Revert "Tweak borland exception design to be more flexible"
This reverts commit 8d186207a042f831c3e7e9cc4b521ffadd7abdeb.

+++ include/cpp98/string.h
@@ -124,7 +124,7 @@ namespace dp {
 		char* errPtr = NULL;
 
 		errno = 0;
-		double val = std::strtod(ptr, &errPtr);
+		float val = std::strtod(ptr, &errPtr);
 		if (errno == ERANGE) {
 			throw std::out_of_range("string conversion out of range");
 		}
@@ -142,7 +142,7 @@ namespace dp {
 		wchar_t* errPtr = NULL;
 
 		errno = 0;
-		double val = std::wcstod(ptr, &errPtr);
+		float val = std::wcstod(ptr, &errPtr);
 		if (errno == ERANGE) {
 			throw std::out_of_range("string conversion out of range");
 		}
